delivery_completion_service/requirements.txt
Flask
google-cloud-firestore==2.3.4